pubsubToSpanner/requirements.txt
google-cloud-spanner==1.4.0
futures==3.2.0; python_version < "3"